feat(home page): implement the home page
implement the home page and the main page
